Fix MacOS build: move tst_lp_dio test to top-level src/test dir
